+++ sqlalchemy_continuum/__init__.py
@@ -17,7 +17,7 @@ from .utils import (
 )
 
 
-__version__ = '1.0-b4'
+__version__ = '1.0b5'
 
 
 versioning_manager = VersioningManager()